backend/requirements.txt
flask==2.0.3
werkzeug==2.0.3
flask-restx==1.3.0
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.46
flask-jwt-extended==4.3.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
flask-migrate==3.1.0
gunicorn==20.1.0